installed Sass and moved some stuff around
