+++ package.json
@@ -6,6 +6,7 @@
     "antd": "^3.24.3",
     "chart.js": "^2.9.1",
     "moment": "^2.24.0",
+    "node-sass": "^4.13.0",
     "pigeon-maps": "^0.14.0",
     "react": "^16.9.0",
     "react-dom": "^16.9.0",
