+++ src/components/DataVisualization/BinMap/index.js
