+++ src/components/DataVisualization/BinTable/index.js
