+++ src/components/Pages/BinStatusPage/index.js
