+++ src/components/Pages/DeviceRegistrationPage/index.js
