+++ src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
+import './css/index.scss';
 import App from './components/App';
 import * as serviceWorker from './serviceWorker';
